requirements.txt
Django==4.2.9
djangorestframework==3.15.2
django-crispy-forms==2.3
crispy-bootstrap4==2024.10
